backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyjwt==2.10.1
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
huggingface-hub==0.19.4
timm==0.9.10
transformers==4.35.2
python-multipart==0.0.6
aiofiles==23.2.1
reportlab==4.0.7
